refactor: rename vendored jsonld-schema to index.ts
Drop the -impl suffix (a Java-ism); the jsonld-schema/ folder already
signals the vendored module, so index.ts reads as its entry point.
